NIAD-3260: When sending a NACK to the sending system, also include the responseText (#1005)
* Refactor: Make test failure messages more readable

Using assertThat combined with contains tells you what the actual string
is and what it was expecting to contain within the failure message.

* When sending a NACK to the sending system, also include responseText

It appears that other GP2GP implementations send the NACK response text
as part of the coded response.

This change brings us in line with these.

Furthermore, one implementation claims that they require this field
to be populated otherwise they will error.