Clear ApplicationContext before each ServiceFailureIT test
It appears that the SpyBeans are causing issues between tests.
One spy can interfer with a subsequent test.

Clearing the ApplicationContext between each test seems to fix this.

Downside of this is that the tests are a bit slower, but given that
the tests already take 5 minutes that isn't a noticable change.