Refactor: Improve clarity of Integration Test failure messages when awaiting for Migration Status (#1022)
* Remove unused method

* Replace isMigrationStatus check with assertThatMigrationStatus

assertThatMigrationStatus provides a much clearer failure message
when the await fails. It provides in the exception message both
the expected condition, and the actual condition, as opposed to
saying "expected true, was false" for isMigrationStatus.

* Improve hasContinueMessageBeenReceived failure message

Use assertThatMigrationStatus as it provides a much clearer message
in the event of a failure. "Expected X, but was Y", as opposed to
"Expected true, but was false". Where X and Y are migration status
values.

* Replace isEhrMigrationCompleted with assertThatMigrationStatus

It gives a clearer assertion message when the assert fails.
"Expected X, actual Y" where X and Y are migration status values
as opposed to "Expected true, got false".